Extract server recovery tracking service
Move recovery task tracking and retry backoff out of the server controller while preserving the public RecoveryTask surface and retry ordering.
