+++ Packages/src/Editor/Infrastructure/Server/UnityCliLoopServerRecoveryTrackingService.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 88018a714b234101b66f17dc78ed67e8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
